+++ crucible-syntax/test/Tests.hs
@@ -80,11 +80,13 @@ testSimulator :: FilePath -> FilePath -> IO ()
 testSimulator inFile outFile =
   do contents <- T.readFile inFile
      withFile outFile WriteMode $ \outh ->
-       simulateProgram inFile contents outh Nothing testOptions
-         (\sym ha ->
-           do os1 <- SyntaxOvrs.setupOverrides sym ha
-              os2 <- TestOvrs.setupOverrides sym ha
-              return $ concat [os1,os2])
+       simulateProgram inFile contents outh Nothing testOptions $
+         defaultSimulateProgramHooks
+           { setupOverridesHook =  \sym ha ->
+               do os1 <- SyntaxOvrs.setupOverrides sym ha
+                  os2 <- TestOvrs.setupOverrides sym ha
+                  return $ concat [os1,os2]
+           }
 
 
 data Lam = Lam [Text] (Datum TrivialAtom) deriving (Eq, Show)