+++ crucible-syntax/CHANGELOG.md
@@ -15,6 +15,27 @@
   Where the `parsedProgGlobals :: Map GlobalName (Pair TypeRepr GlobalVar)` and
   `parsedProgCFGs :: [ACFG ext]` fields of `ParsedProgram` now serve the roles
   previously filled by the first and second fields of the returned tuple.
+* The type of `simulateProgram`'s last argument:
+
+  ```hs
+  simulateProgram
+    :: ...
+    -> (forall p sym ext t st fs. (IsSymInterface sym, sym ~ (ExprBuilder t st fs)) =>
+          sym -> HandleAllocator -> IO [(FnBinding p sym ext,Position)])
+    -> ...
+  ```
+
+  Has changed to the following:
+
+  ```hs
+  simulateProgram
+    :: ...
+    -> SimulateProgramHooks
+    -> ...
+  ```
+
+  Where the `setupOverridesHook` field of `SimulateProgramHooks` now serves the
+  role previously filled by the function argument.
 
 # 0.2